Frontend(components, PairPaycheck): pairable to bills/contributions
